backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
chromadb
pandas
transformers